+++ examples/playground/app.py
@@ -4,7 +4,6 @@ import traceback
 from datetime import timedelta
 
 import pandas as pd
-import spacy
 import streamlit as st
 from output import init_settings as init_output_settings
 from output import scan as scan_output
@@ -12,9 +11,6 @@ from prompt import init_settings as init_prompt_settings
 from prompt import scan as scan_prompt
 
 from llm_guard.vault import Vault
-
-if not spacy.util.is_package("en_core_web_trf"):
-    spacy.cli.download("en_core_web_trf")
 
 PROMPT = "prompt"
 OUTPUT = "output"